requirements.txt
google-auth
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
google-analytics-data==0.18.14
supabase==2.3.0
gotrue==2.8.1
python-dotenv==1.0.0
storage3>=0.7.0
tenacity==9.0.0
openpyxl==3.1.2
psycopg2-binary==2.9.9
